% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocx.R
\name{clean_namespace}
\alias{clean_namespace}
\title{Clean Namespace}
\usage{
clean_namespace(add_file, skeleton_file)
}
\description{
This function cleans the namespace file to remove formats that I don't believe
need to be there.  I am not sure this file is necessary.  Consider deprecating it
if you can confirm that the function does nothing of value.
}
\details{
This function is not meant to be called by the user.
}

