% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocx.R
\name{evaluate_yaml}
\alias{evaluate_yaml}
\title{Evaluate YAML}
\usage{
evaluate_yaml(yfm)
}
\arguments{
\item{yfm}{YAML front matter (stored by knitr as yaml_front_matter)}
}
\value{
The YAML front matter after R code evaluation
}
\description{
This function reads in the YAML variables and evaluates any included R code.
}

