% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocx.R
\name{get_template_info}
\alias{get_template_info}
\title{Get Template Information}
\usage{
get_template_info(yaml_front_matter, reference_docx)
}
\description{
This function will read in the header reference file and replace any codes with
values supplied in the YAML header.
}

