% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocx.R
\name{remove_draft}
\alias{remove_draft}
\title{Remove Draft}
\usage{
remove_draft()
}
\description{
This function will check the header files to determine if the text in the
header files indicate that this is a draft.  If found, the draft text
will be removed.
}

