% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocx.R
\name{rocx}
\alias{rocx}
\title{rocx}
\usage{
rocx(reference_docx, draft = TRUE, keep_old = FALSE, use_bookdown = TRUE,
  ...)
}
\description{
This is the function that is called internally to generate the memos in different
forms.  This is not meant to be called by the user.
}

